reg_water/requirements.txt
requests==2.25.1
apscheduler==3.10.4
pydantic-core==2.41.5
fastapi==0.135.3
uvicorn==0.44.0